+++ test/unix/common.ml
@@ -17,6 +17,7 @@ let random_string () = String.init String_size.length (fun _i -> random_char ())
 
 module Default = struct
   let log_size = 4
+  let lru_size = 0
   let size = 103
 end
 
@@ -24,6 +25,7 @@ module Key = struct
   include Index.Key.String_fixed (String_size)
 
   let v = random_string
+  let pp = Fmt.Dump.string
 end
 
 module Value = struct
@@ -31,6 +33,7 @@ module Value = struct
 
   let v = random_string
   let equal = String.equal
+  let pp = Fmt.Dump.string
 end
 
 type binding = Key.t * Value.t
@@ -105,17 +108,20 @@ struct
 
   let ignore (_ : t) = ()
 
-  let empty_index ?(log_size = Default.log_size) ?flush_callback ?throttle () =
+  let empty_index ?(log_size = Default.log_size) ?(lru_size = Default.lru_size)
+      ?flush_callback ?throttle () =
     let name = fresh_name "empty_index" in
     let cache = Index.empty_cache () in
     let rw =
-      Index.v ?flush_callback ?throttle ~cache ~fresh:true ~log_size name
+      Index.v ?flush_callback ?throttle ~cache ~fresh:true ~log_size ~lru_size
+        name
     in
     let close_all = ref (fun () -> Index.close rw) in
     let tbl = Hashtbl.create 0 in
     let clone ?(fresh = false) ~readonly () =
       let t =
-        Index.v ?flush_callback ?throttle ~cache ~fresh ~log_size ~readonly name
+        Index.v ?flush_callback ?throttle ~cache ~fresh ~log_size ~lru_size
+          ~readonly name
       in
       (close_all := !close_all >> fun () -> Index.close t);
       t
@@ -123,7 +129,8 @@ struct
     { rw; tbl; clone; close_all = (fun () -> !close_all ()) }
 
   let full_index ?(size = Default.size) ?(log_size = Default.log_size)
-      ?(flush_callback = fun () -> ()) ?throttle () =
+      ?(lru_size = Default.lru_size) ?(flush_callback = fun () -> ()) ?throttle
+      () =
     let f =
       (* Disable [flush_callback] while adding initial entries *)
       ref (fun () -> ())
@@ -133,7 +140,7 @@ struct
     let rw =
       Index.v
         ~flush_callback:(fun () -> !f ())
-        ?throttle ~cache ~fresh:true ~log_size name
+        ?throttle ~cache ~fresh:true ~log_size ~lru_size name
     in
     let close_all = ref (fun () -> Index.close rw) in
     let tbl = Hashtbl.create 0 in
@@ -148,7 +155,8 @@ struct
     f := flush_callback (* Enable [flush_callback] *);
     let clone ?(fresh = false) ~readonly () =
       let t =
-        Index.v ~flush_callback ?throttle ~cache ~fresh ~log_size ~readonly name
+        Index.v ~flush_callback ?throttle ~cache ~fresh ~log_size ~lru_size
+          ~readonly name
       in
       (close_all := !close_all >> fun () -> Index.close t);
       t
@@ -160,11 +168,15 @@ struct
     t.close_all ();
     a
 
-  let with_empty_index ?log_size ?flush_callback ?throttle () f =
-    call_then_close (empty_index ?log_size ?flush_callback ?throttle ()) f
+  let with_empty_index ?log_size ?lru_size ?flush_callback ?throttle () f =
+    call_then_close
+      (empty_index ?log_size ?lru_size ?flush_callback ?throttle ())
+      f
 
-  let with_full_index ?log_size ?flush_callback ?throttle ?size () f =
-    call_then_close (full_index ?log_size ?flush_callback ?throttle ?size ()) f
+  let with_full_index ?log_size ?lru_size ?flush_callback ?throttle ?size () f =
+    call_then_close
+      (full_index ?log_size ?lru_size ?flush_callback ?throttle ?size ())
+      f
 end
 
 let ( let* ) f k = f k